fix(sizereport): use commander, fix code
