+++ .gitignore
@@ -1,5 +1,5 @@
 /dist/
-/node_modules/
+**/node_modules/
 .sass-cache/
 /.vscode/
 /.cache/
@@ -16,5 +16,5 @@ yarn.lock
 *.swp
 /workspace/
 /assets/
-/.circleci/css-size-report/package-lock.json
-/.circleci/css-size-report/report.html
+/scripts/sizeReport/package-lock.json
+/scripts/sizeReport/report.html
